backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
SQLAlchemy==2.0.36
werkzeug==3.0.4
python-dotenv==1.0.1
python-docx==1.1.2
Jinja2==3.1.4
email-validator==2.2.0
itsdangerous==2.2.0
pytest==8.3.3
